+++ apps/api/src/Application/CommandParsers/NmapParser.cs
@@ -3,9 +3,9 @@ using api_v2.Application.Services;
 using api_v2.Controllers;
 using api_v2.Domain.Entities;
 
-namespace api_v2.Application.CommandProcessors;
+namespace api_v2.Application.CommandParsers;
 
-public class NmapProcessor(IAttachmentStorage attachmentStorage) : IProcessor
+public class NmapParser(IAttachmentStorage attachmentStorage) : ICommandParser
 {
     private readonly AttachmentFilePath _attachmentFilePath = new();
 
@@ -14,9 +14,9 @@ public class NmapProcessor(IAttachmentStorage attachmentStorage) : IProcessor
     public string ExternalUrl => "https://nmap.org/";
     public bool IsConfigured => true;
 
-    public ProcessorResult Process(CommandProcessorJob job)
+    public CommandParserResult Parse(CommandProcessorJob job)
     {
-        var result = new ProcessorResult();
+        var result = new CommandParserResult();
 
         using var stream = attachmentStorage.GetFileStreamAsync(job.FilePath).GetAwaiter().GetResult();
         XDocument xml;